+++ .devcontainer/python-3.9/devcontainer.json
@@ -6,7 +6,7 @@
     "service": "python-3.9",
     "forwardPorts": [],
     "shutdownAction": "stopCompose",
-    "workspaceFolder": "/workspaces/weatherman",
+    "workspaceFolder": "/workspaces/weatherapi",
     "remoteEnv": {},
     "features": {
         "ghcr.io/devcontainers/features/common-utils:2": {
