+++ .github/ISSUE_TEMPLATE/config.yml
@@ -5,4 +5,4 @@ contact_links:
     about: Please report security vulnerabilities to info@airt.ai
   - name: Question or Problem
     about: Ask a question or ask about a problem in GitHub Discussions.
-    url: https://github.com/airtai/weatherman/discussions/categories/questions
+    url: https://github.com/airtai/weatherapi/discussions/categories/questions
