+++ .github/workflows/docker_cleanup.yml
@@ -16,7 +16,7 @@ jobs:
         with:
           # NOTE: at now only orgs is supported
           owner: airtai
-          name: weatherman
+          name: weatherapi
 
           token: ${{ secrets.GITHUB_TOKEN }}
           # Keep latest N untagged images
