+++ .github/workflows/test.yaml
@@ -29,11 +29,11 @@ jobs:
           pip install -e ".[docs,lint]"
       - name: Run mypy
         shell: bash
-        run: mypy weatherman tests
+        run: mypy weatherapi tests
 
       - name: Run bandit
         shell: bash
-        run: bandit -c pyproject.toml -r weatherman
+        run: bandit -c pyproject.toml -r weatherapi
 
       - name: Run Semgrep
         shell: bash
@@ -146,7 +146,7 @@ jobs:
       - run: ls -la coverage
       - run: coverage combine coverage
       - run: coverage report
-      - run: coverage html --show-contexts --title "weatherman coverage for ${{ github.sha }}"
+      - run: coverage html --show-contexts --title "weatherapi coverage for ${{ github.sha }}"
 
       - name: Store coverage html
         uses: actions/upload-artifact@v4
