+++ scripts/lint.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 echo "Running pyup_dirs..."
-pyup_dirs --py38-plus --recursive weatherman tests
+pyup_dirs --py38-plus --recursive weatherapi tests
 
 echo "Running ruff linter (isort, flake, pyupgrade, etc. replacement)..."
 ruff check
@@ -10,4 +10,4 @@ echo "Running ruff formater (black replacement)..."
 ruff format
 
 # echo "Running black..."
-# black weatherman examples tests docs
+# black weatherapi examples tests docs
