+++ scripts/run_server.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-uvicorn weatherman.app:app --workers 2 --host 0.0.0.0 --proxy-headers
+uvicorn weatherapi.app:app --workers 2 --host 0.0.0.0 --proxy-headers
