+++ weatherapi/__about__.py
@@ -1,0 +1,3 @@
+"""A simple weather API made for usage with FastAgency."""
+
+__version__ = "0.0.0dev0"
