+++ weatherapi/__init__.py
@@ -1,0 +1,3 @@
+"""A simple weather API made for usage with FastAgency."""
+
+from .__about__ import __version__  # noqa: F401
