--- weatherman/__about__.py
@@ -1,3 +1,0 @@
-"""An API to use with FastAgency."""
-
-__version__ = "0.0.0dev0"
