--- weatherman/__init__.py
@@ -1,3 +1,0 @@
-"""An API to use with FastAgency."""
-
-from .__about__ import __version__  # noqa: F401
